requirements/required.txt
# setup
setuptools==63.1.0

# install
einops==0.4.1
fiona==1.9a2
kornia==0.6.5
matplotlib==3.5.2
numpy==1.23.0;python_version>='3.8'
omegaconf==2.2.2
packaging==21.3
pillow==9.2.0
protobuf==3.20.1
pyproj==3.3.1;python_version>='3.8'
pytorch-lightning==1.6.4
rasterio==1.3b3;python_version>='3.8'
rtree==1.0.0
scikit-learn==1.1.1;python_version>='3.8'
segmentation-models-pytorch==0.2.1
shapely==1.8.2
timm==0.4.12
torch==1.12.0
torchmetrics==0.9.2
torchvision==0.13.0
